requirements.txt
python-dotenv==1.0.1
langchain==0.2.3
langchain-community==0.2.4
neo4j==5.20.0


# aiohttp==3.9.5
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.4.0
# attrs==23.2.0
# certifi==2024.6.2
# charset-normalizer==3.3.2
# dataclasses-json==0.6.6
# distro==1.9.0
# frozenlist==1.4.1
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.0
# idna==3.7
# iniconfig==2.0.0
# jsonpatch==1.33
# jsonpointer==2.4
# langchain-core==0.2.5
# langchain-openai==0.0.2
# langchain-text-splitters==0.2.1
# langchainhub==0.1.14
# langsmith==0.1.75
# marshmallow==3.21.3
# # Editable Git install with no remote (movie_chatbot==0.1)
# -e /Users/mike/code/movie_chatbot
# multidict==6.0.5
# mypy-extensions==1.0.0
# neo4j==5.20.0
# numpy==1.26.4
# openai==1.7.2
# orjson==3.10.3
# packaging==23.2
# pluggy==1.5.0
# pydantic==2.7.3
# pydantic_core==2.18.4
# pytest==8.2.2
# pytz==2024.1
# PyYAML==6.0.1
# regex==2024.5.15
# requests==2.32.3
# setuptools==70.0.0
# sniffio==1.3.1
# SQLAlchemy==2.0.30
# tenacity==8.3.0
# tiktoken==0.5.2
# tqdm==4.66.4
# types-requests==2.32.0.20240602
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# urllib3==2.2.1
# yarl==1.9.4
